requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
opencv-python>=4.9.0
mediapipe>=0.10.9
numpy>=1.26.0
